GameDIsplay() 85% complete.
Committing to master because my Git still won’t let me work in branch
and do a pull request.

Had to rewrite to get it to display properly.
updateDisplay() iterates over input grid and displays the appropriate
image.
Fixed the frame so it is not resizable.

Also added space background, default image for sassyvader space,
buttons (quit even terminates the program, this can be changed later).
Preview also works (Propose change input to an 8x8 boolean[][] grid)

I think it might be a little small, but we can fix that later. This is
now ready for integration with System and GameRunner.

Next Steps:
I have a decent idea of how to add popup windows for game over and
pause.
May want to resize everything.
The left margin is small and bugging me but haven’t been able to figure
out how to fix it.